:heavy_plus_sign: Add ProjectInfo class
